fixing front

+++ exams supervisor/exams supervisor/Model/TeacherInfo.cs
@@ -7,7 +7,7 @@ namespace exams_supervisor.Model
 		
 		public required string fullName { get; set; }
 		public required string grade { get; set; }
-		public int hourlyLoad  { get; set; }
+		public double hourlyLoad  { get; set; }
 
 	}
 }
